Finished JSA Form and added date picker.
Finished bottom section of the JSA front end. Added date-picker to the JSA form and the daily job report form. Although it does not currently work. I think we need to tweak the date picker a little. When onChange is called it can't handle the date value. Will continue to look into it.